change type var float to long double for amount

+++ ex00/BitcoinExchange.hpp
@@ -6,7 +6,7 @@
 /*   By: bcarolle <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/17 18:07:40 by bcarolle          #+#    #+#             */
-/*   Updated: 2024/05/27 23:28:44 by bcarolle         ###   ########.fr       */
+/*   Updated: 2024/06/02 19:14:34 by bcarolle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,7 @@
 # include <string>
 # include <stdlib.h>
 # include <sstream>
+# include <iomanip>
 
 class BitcoinExchange
 {
@@ -36,7 +37,7 @@ class BitcoinExchange
 				virtual const char* what() const throw();
 		};
 	private:
-		void	findData(std::string date, float amount);
+		void	findData(std::string date, long double amount);
 		std::map<std::string, float> _data;
 		std::map<std::string, float> _input;
 };
